requirements.txt
pycryptodome==3.20.0
flask==3.0.3
Flask-SQLAlchemy==3.1.1
